Moved room count fetch to global state
