+++ src/GlobalStateContext.tsx
@@ -9,6 +9,8 @@ import { Session } from "./types/session";
 import { TalkComments } from "./types/talkComment";
 import { NoteItemProps } from "./components/NoteItem/NoteItem";
 import { createMonogram } from "./utils/helpers";
+import { RoomWithUserCounts } from "./types/room";
+import { CATEGORIES, CATEGORY_NAMES_TO_ID_MAP } from "./utils/constants";
 
 interface GlobalState {
   username: string;
@@ -37,6 +39,8 @@ interface GlobalState {
   setNotes: React.Dispatch<React.SetStateAction<NoteItemProps[]>>;
   talkActivity: Map<string, number>;
   setTalkActivity: React.Dispatch<React.SetStateAction<Map<string, number>>>;
+  orderedList: RoomWithUserCounts[];
+  setOrderedList: React.Dispatch<React.SetStateAction<RoomWithUserCounts[]>>;
 }
 
 const GlobalStateContext = createContext<GlobalState | undefined>(undefined);
@@ -83,6 +87,16 @@ export const GlobalStateProvider: React.FC<GlobalStateProviderProps> = ({
 
   const [talkActivity, setTalkActivity] = useState<Map<string, number>>(
     new Map<string, number>()
+  );
+
+  // List of room, ordered by user count
+  const [orderedList, setOrderedList] = useState<RoomWithUserCounts[]>(
+    CATEGORIES.map((catName) => ({
+      topic: CATEGORY_NAMES_TO_ID_MAP.get(catName),
+      url: "null",
+      gateway: "null",
+      userCount: undefined,
+    }))
   );
 
   useEffect(() => {
@@ -142,6 +156,8 @@ export const GlobalStateProvider: React.FC<GlobalStateProviderProps> = ({
         setNotes,
         talkActivity,
         setTalkActivity,
+        orderedList,
+        setOrderedList
       }}
     >
       {children}
