+++ src/MainRouter.tsx
@@ -47,6 +47,7 @@ import {
   isUserRegistered,
   getPrivateKey,
 } from "./utils/helpers";
+import { RoomWithUserCounts } from "./types/room";
 
 const MainRouter = (): ReactElement => {
   const {
@@ -64,6 +65,8 @@ const MainRouter = (): ReactElement => {
     notes,
     setNotes,
     setTalkActivity,
+    orderedList,
+    setOrderedList
   } = useGlobalState();
   const [sessionsReference, setSessionsReference] = useState<string>("");
   const [isBeeRunning, setBeeRunning] = useState<boolean>(false);
@@ -298,8 +301,27 @@ const MainRouter = (): ReactElement => {
     }
   };
 
+  // User count refreshes every 15 minutes on backend. With this function, we fetch the stored values.
+  const fetchUserCount = async () => {
+    const roomsWithUserCount: RoomWithUserCounts[] = await fetch(
+      process.env.BACKEND_API_URL + "/user-count"
+    )
+      .then((res) => res.json())
+      .catch((err) => console.error("Error fetching user counts ", err));
+
+    if (roomsWithUserCount !== undefined) {
+      const orderedRooms = roomsWithUserCount.sort(
+        (a, b) => b.userCount! - a.userCount!
+      );
+      setOrderedList(orderedRooms);
+    }
+
+    console.log("Rooms with user counts: ", orderedList);
+  };
+
   useEffect(() => {
     preLoadTalks();
+    fetchUserCount();
   }, [recentSessions]);
 
   const calcActivity = () => {
